Enhancement: Extract StringContainsTest
