+++ tests/Framework/ConstraintTest.php
@@ -2125,86 +2125,6 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintStringContains()
-    {
-        $constraint = Assert::stringContains('foo');
-
-        $this->assertFalse($constraint->evaluate('barbazbar', '', true));
-        $this->assertTrue($constraint->evaluate('barfoobar', '', true));
-        $this->assertEquals('contains "foo"', $constraint->toString());
-        $this->assertCount(1, $constraint);
-
-        try {
-            $constraint->evaluate('barbazbar');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-Failed asserting that 'barbazbar' contains "foo".
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintStringContainsWhenIgnoreCase()
-    {
-        $constraint = Assert::stringContains('oryginał', true);
-
-        $this->assertFalse($constraint->evaluate('oryginal', '', true));
-        $this->assertTrue($constraint->evaluate('ORYGINAŁ', '', true));
-        $this->assertTrue($constraint->evaluate('oryginał', '', true));
-        $this->assertEquals('contains "oryginał"', $constraint->toString());
-        $this->assertEquals(1, \count($constraint));
-
-        $this->expectException(ExpectationFailedException::class);
-
-        $constraint->evaluate('oryginal');
-    }
-
-    public function testConstraintStringContainsForUtf8StringWhenNotIgnoreCase()
-    {
-        $constraint = Assert::stringContains('oryginał', false);
-
-        $this->assertFalse($constraint->evaluate('oryginal', '', true));
-        $this->assertFalse($constraint->evaluate('ORYGINAŁ', '', true));
-        $this->assertTrue($constraint->evaluate('oryginał', '', true));
-        $this->assertEquals('contains "oryginał"', $constraint->toString());
-        $this->assertEquals(1, \count($constraint));
-
-        $this->expectException(ExpectationFailedException::class);
-
-        $constraint->evaluate('oryginal');
-    }
-
-    public function testConstraintStringContains2()
-    {
-        $constraint = Assert::stringContains('foo');
-
-        try {
-            $constraint->evaluate('barbazbar', 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-custom message
-Failed asserting that 'barbazbar' contains "foo".
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
     public function testConstraintStringNotContains()
     {
         $constraint = Assert::logicalNot(
